moved TextPopout script, fixed namespace
